feat(frontend): Add file picker to choose the working directory #221
Added a file picker feature to top of the chat window.

Now a user can select the directory to work with.

Once a directory is chosen the user has the  option to edit the directory also.(i.e choose another directory)

I require backend help to initialise the model within the directory.