feat: Expose session_id to sandbox/runtime container
add the envvar `OPENHANDS_SESSION_ID` to the list of default environment variables populated within the runtime environment. 

This enables some tooling to be able to discern the notion of `what session am I` which, if combined with the exposure of the settings store (not quite as optimal, but much easier to accomplish) to the sandbox environment allows for programatic evaluation of the session-scoped `metadata.json` file located in `settings/sessions/<OPENHANDS_SESSION_ID>/` as `settings/sessions/<OPENHANDS_SESSION_ID>/metadata.json`

IMO a better solution would be to deposit the `metadata.json`  in /tmp or something similar which would facilitate the setting of additional vars like `CODEHOST_ORGANIZATION` `CODEHOST_PROJECT_PATH` `CODEHOST_SELECTED_PROJECT` `CODEHOST_PROJECT_ACTIVE_BRANCH`

subsequently, and with a little voodoo / interaction with the user, this allows for the determination of  `PROJECT_DEFAULT_BRANCH` `PROJECT_WORKING_BRANCH` `PROJECT_TARGET_BRANCH` 
(or something)... but that's not really relevant to this change.